Add watchOS, tvOS and macOS support
I'm in the process of splitting the code of the Pyto app with its runtime, something like what ios_system is to a-Shell. And I want to be able to cross compile that to as many platforms as possible with all the Python packages, LLVM and other commands compiled primarily for iOS and ios_system. We could then just add a Swift package that will include all the binaries as libraries on any platform. So I guess this project makes sense for more than just iOS. It could be part of a multi platform SDK where we have the same versions of everything and it can run the same way on any platform without depending on system libraries.